backend/requirements.txt
# Core Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1

# Database
motor==3.3.2
pymongo==4.6.1

# Caching
redis==5.0.1
aioredis==2.0.1

# Authentication & Security
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
cryptography==42.0.2

# Validation & Settings
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0

# File Handling & Utils
python-multipart==0.0.9
python-dotenv==1.0.0

# CORS
starlette==0.36.3

# Async Support
anyio==4.2.0

# AI Integration
#open-interpreter==0.4.0
#pip install open-interpreter==0.3.14 